+++ crates/module-system/sov-modules-api/src/transaction/mod.rs
@@ -8,8 +8,6 @@ use borsh::{BorshDeserialize, BorshSerialize};
 pub use data::{AuthenticatedTransactionData, Credentials, PriorityFeeBips, TxDetails};
 pub(crate) use rewards::transaction_consumption_helper;
 pub use rewards::{ProverReward, RemainingFunds, SequencerReward, TransactionConsumption};
-use serde::{Deserialize, Serialize};
-use sov_rollup_interface::common::SafeVec;
 #[cfg(feature = "native")]
 pub use sov_rollup_interface::crypto::PrivateKey;
 use sov_rollup_interface::crypto::{SigVerificationError, Signature};
@@ -18,19 +16,19 @@ use sov_rollup_interface::zk::CryptoSpec;
 use sov_rollup_interface::TxHash;
 use sov_universal_wallet::schema::UniversalWallet;
 use thiserror::Error;
+pub use types::{v0::Version0, v1::Version1};
+pub use unsigned::UnsignedTransaction;
 
 use crate::capabilities::UniquenessData;
 use crate::{
-    Amount, CryptoSpecExt, DispatchCall, Gas, GasMeter, GasMeteringError, GasSpec,
-    MeteredBorshDeserialize, MeteredBorshDeserializeError, MeteredSigVerificationError,
-    MeteredSignature, Spec,
+    CryptoSpecExt, DispatchCall, Gas, GasMeter, GasMeteringError, GasSpec, MeteredBorshDeserialize,
+    MeteredBorshDeserializeError, MeteredSigVerificationError, MeteredSignature, Spec,
 };
 
 #[cfg(test)]
 mod tests;
-
-/// The maximum number of signers allowed in a multisig.
-pub const MAX_SIGNERS: usize = 21;
+mod types;
+mod unsigned;
 
 /// Structures that implement this trait represent a call message that can be included in a
 /// transaction.
@@ -43,37 +41,6 @@ pub trait TransactionCallable {
 
 impl<D: DispatchCall> TransactionCallable for D {
     type Call = D::Decodable;
-}
-
-#[derive(
-    derive_more::Debug,
-    Clone,
-    borsh::BorshDeserialize,
-    serde::Serialize,
-    serde::Deserialize,
-    borsh::BorshSerialize,
-    UniversalWallet,
-)]
-#[serde(bound = "Call: serde::Serialize + serde::de::DeserializeOwned")]
-/// V0 transaction.
-pub struct Version0<Call, S: Spec, C: CryptoSpecExt = <S as Spec>::CryptoSpec> {
-    /// The signature of the transaction.
-    #[serde(with = "hex_field_format")]
-    #[sov_wallet(display = "hex")]
-    pub signature: C::Signature,
-    /// The public key of the sender of the transaction.
-    #[serde(with = "hex_field_format")]
-    #[sov_wallet(display = "hex")]
-    pub pub_key: C::PublicKey,
-    /// The runtime call of the transaction.
-    #[sov_wallet(
-        bound = "Call: sov_rollup_interface::sov_universal_wallet::schema::UniversalWallet"
-    )]
-    pub runtime_call: Call,
-    /// Uniqueness identifier of this transaction. see [`UniquenessData`] for more details.
-    pub uniqueness: UniquenessData,
-    /// The transaction metadata. Contains gas parameters and the chain ID.
-    pub details: TxDetails<S>,
 }
 
 #[derive(
@@ -103,112 +70,6 @@ impl<C: CryptoSpecExt> PubKeyAndSignature<C> {
     }
 }
 
-#[derive(
-    derive_more::Debug,
-    Clone,
-    borsh::BorshDeserialize,
-    serde::Serialize,
-    serde::Deserialize,
-    borsh::BorshSerialize,
-    UniversalWallet,
-)]
-#[serde(bound = "Call: serde::Serialize + serde::de::DeserializeOwned")]
-/// A V1 (multisig) transaction. The number of signers is capped at 10.
-///
-/// The credential ID for a multisig is hash(borsh(min_signers) || borsh(sort(pub_keys)))
-pub struct Version1<Call, S: Spec, C: CryptoSpecExt = <S as Spec>::CryptoSpec> {
-    /// The signatures of the transaction.
-    #[borsh(bound(
-        serialize = "PubKeyAndSignature<C>: BorshSerialize",
-        deserialize = "PubKeyAndSignature<C>: BorshDeserialize",
-    ))]
-    pub signatures: SafeVec<PubKeyAndSignature<C>, MAX_SIGNERS>,
-    /// The credential IDs that are part of the multisig but not used to sign the transaction.
-    // This is used to compute the credential ID statelessly
-    #[borsh(bound(
-        serialize = "<C as CryptoSpec>::PublicKey: BorshSerialize",
-        deserialize = "<C as CryptoSpec>::PublicKey: BorshDeserialize",
-    ))]
-    pub unused_pub_keys: SafeVec<C::PublicKey, MAX_SIGNERS>,
-    /// The minimum number of signers required to sign the transaction. In a 3/5 multisig, this would be 3.
-    /// Note that...
-    ///  - The transaction will be valid if at least `min_signers` signers have signed the transaction, but more signers are allowed. (This is useful for record keeping in case of, say, a unanimous decision)
-    ///  - If an invalid signature is detected, the transaction will be invalid. This allows us to use batch verification.
-    // This is used to compute the credential ID statelessly.
-    pub min_signers: u8,
-    /// The runtime call of the transaction.
-    #[sov_wallet(
-        bound = "Call: sov_rollup_interface::sov_universal_wallet::schema::UniversalWallet"
-    )]
-    pub runtime_call: Call,
-    /// Uniqueness identifier of this transaction. see [`UniquenessData`] for more details.
-    pub uniqueness: UniquenessData,
-    /// The transaction metadata. Contains gas parameters and the chain ID.
-    pub details: TxDetails<S>,
-}
-
-impl<Call: BorshSerialize, S: Spec, C: CryptoSpecExt> Version1<Call, S, C> {
-    /// Signs the transaction with the given key but does not add the signature to the list in the transaction.
-    #[cfg(feature = "native")]
-    pub fn sign_without_adding(&self, key: &C::PrivateKey, chain_hash: &[u8; 32]) -> C::Signature {
-        key.sign(&self.serialize_for_signing(chain_hash))
-    }
-
-    /// Signs and adds the signature to the transaction.
-    #[cfg(feature = "native")]
-    pub fn sign(&mut self, key: &C::PrivateKey, chain_hash: &[u8; 32]) -> anyhow::Result<()> {
-        let signature = self.sign_without_adding(key, chain_hash);
-        self.add_signature(signature, key.pub_key())
-    }
-
-    /// Serializes only the `UnsignedTransaction` part of the transaction and appens the chain_hash
-    pub fn serialize_for_signing(&self, chain_hash: &[u8; 32]) -> Vec<u8> {
-        let mut out = Vec::with_capacity(64); // Preallocate a little capacity to avoid excessive reallocations
-        BorshSerialize::serialize(&self.runtime_call, &mut out)
-            .expect("Serialization to vec is infallible");
-        BorshSerialize::serialize(&self.uniqueness, &mut out)
-            .expect("Serialization to vec is infallible");
-        BorshSerialize::serialize(&self.details, &mut out)
-            .expect("Serialization to vec is infallible");
-        out.extend_from_slice(chain_hash);
-        out
-    }
-
-    /// Adds a signature to the signing set of the multisig, removing the public key from the set of unused pub keys.
-    pub fn add_signature(
-        &mut self,
-        signature: C::Signature,
-        pub_key: C::PublicKey,
-    ) -> anyhow::Result<()> {
-        use crate::prelude::anyhow::Context;
-        if let Some(index) = self.unused_pub_keys.iter().position(|k| k == &pub_key) {
-            self.signatures
-                .try_push(PubKeyAndSignature { signature, pub_key })
-                .context("Too many signatures in multisig")?;
-            self.unused_pub_keys.remove(index);
-            Ok(())
-        } else {
-            anyhow::bail!("Public key is not a member of the multisig or has already signed");
-        }
-    }
-
-    /// Checks if enough signers have signed the transaction for it to be valid.
-    pub fn is_fully_signed(&self) -> bool {
-        self.signatures.len() >= self.min_signers as usize
-    }
-}
-
-impl<R: TransactionCallable, S: Spec> From<Version0<R::Call, S>> for Transaction<R, S> {
-    fn from(value: Version0<R::Call, S>) -> Self {
-        Transaction::V0(value)
-    }
-}
-
-impl<R: TransactionCallable, S: Spec> From<Version1<R::Call, S>> for Transaction<R, S> {
-    fn from(value: Version1<R::Call, S>) -> Self {
-        Transaction::V1(value)
-    }
-}
 #[derive(
     derive_more::Debug, // derive_more uses the correct bound of TransactionCallable::RuntimeCall
     Clone,
@@ -496,117 +357,6 @@ impl<R: TransactionCallable, S: Spec, C: CryptoSpecExt> Transaction<R, S, C> {
     }
 }
 
-/// An unsent transaction with the required data to be submitted to the DA layer
-#[derive(
-    derive_more::Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, UniversalWallet,
-)]
-#[serde(bound = "R::Call: serde::Serialize + serde::de::DeserializeOwned")]
-pub struct UnsignedTransaction<R: TransactionCallable, S: Spec> {
-    /// The runtime call
-    pub runtime_call: R::Call,
-    /// The uniqueness identifier
-    pub uniqueness: UniquenessData,
-    /// Data related to fees and gas handling.
-    pub details: TxDetails<S>,
-}
-// Manually implemented to ensure correct trait bounds for the same reason as for `Transaction`
-// above
-impl<R: TransactionCallable, S: Spec> PartialEq for UnsignedTransaction<R, S> {
-    fn eq(&self, other: &Self) -> bool {
-        self.runtime_call == other.runtime_call
-            && self.uniqueness == other.uniqueness
-            && self.details == other.details
-    }
-}
-impl<R: TransactionCallable, S: Spec> Eq for UnsignedTransaction<R, S> {}
-
-#[cfg(feature = "native")]
-impl<R: TransactionCallable, S: Spec> UnsignedTransaction<R, S> {
-    /// Signs the [`UnsignedTransaction`] and returns the resulting [`Transaction`].
-    pub fn sign(
-        self,
-        private_key: &<S::CryptoSpec as CryptoSpec>::PrivateKey,
-        chain_hash: &[u8; 32],
-    ) -> Transaction<R, S> {
-        Transaction::new_signed_tx(private_key, chain_hash, self)
-    }
-}
-
-impl<R: TransactionCallable, S: Spec> UnsignedTransaction<R, S> {
-    /// Creates a new [`UnsignedTransaction`] with the given arguments.
-    pub const fn new(
-        runtime_call: R::Call,
-        chain_id: u64,
-        max_priority_fee_bips: PriorityFeeBips,
-        max_fee: Amount,
-        uniqueness: UniquenessData,
-        gas_limit: Option<S::Gas>,
-    ) -> Self {
-        Self {
-            runtime_call,
-            uniqueness,
-            details: TxDetails {
-                max_priority_fee_bips,
-                max_fee,
-                gas_limit,
-                chain_id,
-            },
-        }
-    }
-
-    /// Creates a new unsigned transaction with the provided metadata.
-    pub const fn new_with_details(
-        runtime_call: R::Call,
-        uniqueness: UniquenessData,
-        details: TxDetails<S>,
-    ) -> Self {
-        Self {
-            runtime_call,
-            uniqueness,
-            details,
-        }
-    }
-
-    /// Creates a new [`Transaction`] from this [`UnsignedTransaction`] when given a signature
-    /// and a public key.
-    pub fn to_signed_tx<C: CryptoSpecExt>(
-        self,
-        pub_key: C::PublicKey,
-        signature: C::Signature,
-    ) -> Transaction<R, S, C> {
-        Transaction::new_with_details_v0(
-            pub_key,
-            self.runtime_call,
-            signature,
-            self.uniqueness,
-            self.details,
-        )
-    }
-
-    /// Creates a new `V1` transaction from this unsigned transaction.
-    pub fn to_multisig_tx(
-        self,
-        multisig: Multisig<<S::CryptoSpec as CryptoSpec>::PublicKey>,
-    ) -> Version1<R::Call, S> {
-        Version1 {
-            signatures: SafeVec::new(),
-            unused_pub_keys: multisig
-                .signers
-                .try_into()
-                .expect("Too many signers in multisig"),
-            min_signers: multisig.required_signers,
-            runtime_call: self.runtime_call,
-            uniqueness: self.uniqueness,
-            details: self.details,
-        }
-    }
-
-    /// Returns a copy of the RuntimeCall from this unsigned transaction.
-    pub fn call(&self) -> R::Call {
-        self.runtime_call.clone()
-    }
-}
-
 /// A struct containing an authenticated transaction and its associated hash.
 pub struct AuthenticatedTransactionAndRawHash<S: Spec> {
     /// Hash of raw bytes.
@@ -615,7 +365,7 @@ pub struct AuthenticatedTransactionAndRawHash<S: Spec> {
     pub authenticated_tx: AuthenticatedTransactionData<S>,
 }
 
-mod hex_field_format {
+pub(crate) mod hex_field_format {
     use serde::{Deserialize, Deserializer, Serializer};
 
     pub fn serialize<S, V>(value: &V, serializer: S) -> Result<S::Ok, S::Error>
